+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 4.2 (unreleased)
 ----------------
 
-- Add support for Python 3.8.
+- Add support for Python 3.8, 3.9 and 3.10.
 
 - Drop support for Python 3.4.